// reactfpv1/src/interfaces/users/UnnormalaziedUser.ts
export interface unnormalizedUser {
  firstName: string;
  middle?: string;
  lastName: string;
  phone: string;
  email: string;
  password: string;
  image?: string;
  alt?: string;
  state?: string;
  country: string;
  city: string;
  street: string;
  houseNumber: number;
  zip: number;
  isBusiness: boolean;
}